wrapper: extract service implementation to to own module
